Oregon Update.

+++ src/scripts/World/npc_professions.cpp
@@ -406,7 +406,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             break;
             //Learn Alchemy
         case GOSSIP_ACTION_INFO_DEF + 1:
-            if (!pPlayer->HasSpell(S_TRANSMUTE) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_TRANSMUTE) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_TRANSMUTE, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -415,7 +415,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 2:
-            if (!pPlayer->HasSpell(S_ELIXIR) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_ELIXIR) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_ELIXIR, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -424,7 +424,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 3:
-            if (!pPlayer->HasSpell(S_POTION) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_POTION) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_POTION, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -434,7 +434,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             break;
             //Unlearn Alchemy
         case GOSSIP_ACTION_INFO_DEF + 4:
-            if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+            if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
             {
                 pCreature->CastSpell(pPlayer, S_UNLEARN_TRANSMUTE, true);
                 pPlayer->ModifyMoney(-DoHighUnlearnCost(pPlayer));
@@ -443,7 +443,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 5:
-            if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+            if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
             {
                 pCreature->CastSpell(pPlayer, S_UNLEARN_ELIXIR, true);
                 pPlayer->ModifyMoney(-DoHighUnlearnCost(pPlayer));
@@ -452,7 +452,7 @@ void SendActionMenu_npc_prof_alchemy(Player* pPlayer, Creature* pCreature, uint3
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 6:
-            if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+            if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
             {
                 pCreature->CastSpell(pPlayer, S_UNLEARN_POTION, true);
                 pPlayer->ModifyMoney(-DoHighUnlearnCost(pPlayer));
@@ -638,7 +638,7 @@ void SendActionMenu_npc_prof_blacksmith(Player* pPlayer, Creature* pCreature, ui
             }
             else if (EquippedOk(pPlayer,S_UNLEARN_WEAPON))
             {
-                if (pPlayer->GetMoney() >= DoLowUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoLowUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_WEAPON, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_WEAPON);
@@ -657,7 +657,7 @@ void SendActionMenu_npc_prof_blacksmith(Player* pPlayer, Creature* pCreature, ui
         case GOSSIP_ACTION_INFO_DEF + 4:
             if (EquippedOk(pPlayer,S_UNLEARN_ARMOR))
             {
-                if (pPlayer->GetMoney() >= DoLowUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoLowUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_ARMOR, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_ARMOR);
@@ -686,7 +686,7 @@ void SendActionMenu_npc_prof_blacksmith(Player* pPlayer, Creature* pCreature, ui
         case GOSSIP_ACTION_INFO_DEF + 8:
             if (EquippedOk(pPlayer,S_UNLEARN_HAMMER))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_HAMMER, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_HAMMER);
@@ -700,7 +700,7 @@ void SendActionMenu_npc_prof_blacksmith(Player* pPlayer, Creature* pCreature, ui
         case GOSSIP_ACTION_INFO_DEF + 9:
             if (EquippedOk(pPlayer,S_UNLEARN_AXE))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_AXE, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_AXE);
@@ -714,7 +714,7 @@ void SendActionMenu_npc_prof_blacksmith(Player* pPlayer, Creature* pCreature, ui
         case GOSSIP_ACTION_INFO_DEF + 10:
             if (EquippedOk(pPlayer,S_UNLEARN_SWORD))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_SWORD, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_SWORD);
@@ -867,7 +867,7 @@ void SendActionMenu_npc_prof_leather(Player* pPlayer, Creature* pCreature, uint3
         case GOSSIP_ACTION_INFO_DEF + 1:
             if (EquippedOk(pPlayer,S_UNLEARN_DRAGON))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_DRAGON, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_DRAGON);
@@ -881,7 +881,7 @@ void SendActionMenu_npc_prof_leather(Player* pPlayer, Creature* pCreature, uint3
         case GOSSIP_ACTION_INFO_DEF + 2:
             if (EquippedOk(pPlayer,S_UNLEARN_ELEMENTAL))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_ELEMENTAL, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_ELEMENTAL);
@@ -895,7 +895,7 @@ void SendActionMenu_npc_prof_leather(Player* pPlayer, Creature* pCreature, uint3
         case GOSSIP_ACTION_INFO_DEF + 3:
             if (EquippedOk(pPlayer,S_UNLEARN_TRIBAL))
             {
-                if (pPlayer->GetMoney() >= DoMedUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoMedUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_TRIBAL, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_TRIBAL);
@@ -1015,7 +1015,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
             break;
             //Learn Tailor
         case GOSSIP_ACTION_INFO_DEF + 1:
-            if (!pPlayer->HasSpell(S_SPELLFIRE) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_SPELLFIRE) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_SPELLFIRE, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -1024,7 +1024,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 2:
-            if (!pPlayer->HasSpell(S_MOONCLOTH) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_MOONCLOTH) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_MOONCLOTH, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -1033,7 +1033,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
             pPlayer->CLOSE_GOSSIP_MENU();
             break;
         case GOSSIP_ACTION_INFO_DEF + 3:
-            if (!pPlayer->HasSpell(S_SHADOWEAVE) && pPlayer->GetMoney() >= DoLearnCost(pPlayer))
+            if (!pPlayer->HasSpell(S_SHADOWEAVE) && pPlayer->GetMoney() >= uint32(DoLearnCost(pPlayer)))
             {
                 pPlayer->CastSpell(pPlayer, S_LEARN_SHADOWEAVE, true);
                 pPlayer->ModifyMoney(-DoLearnCost(pPlayer));
@@ -1045,7 +1045,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
         case GOSSIP_ACTION_INFO_DEF + 4:
             if (EquippedOk(pPlayer,S_UNLEARN_SPELLFIRE))
             {
-                if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_SPELLFIRE, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_SPELLFIRE);
@@ -1059,7 +1059,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
         case GOSSIP_ACTION_INFO_DEF + 5:
             if (EquippedOk(pPlayer,S_UNLEARN_MOONCLOTH))
             {
-                if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_MOONCLOTH, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_MOONCLOTH);
@@ -1073,7 +1073,7 @@ void SendActionMenu_npc_prof_tailor(Player* pPlayer, Creature* pCreature, uint32
         case GOSSIP_ACTION_INFO_DEF + 6:
             if (EquippedOk(pPlayer,S_UNLEARN_SHADOWEAVE))
             {
-                if (pPlayer->GetMoney() >= DoHighUnlearnCost(pPlayer))
+                if (pPlayer->GetMoney() >= uint32(DoHighUnlearnCost(pPlayer)))
                 {
                     pPlayer->CastSpell(pPlayer, S_UNLEARN_SHADOWEAVE, true);
                     ProfessionUnlearnSpells(pPlayer, S_UNLEARN_SHADOWEAVE);
